Move class name resolver to a different namespace
